+++ tutorials/point_geometry/point_geometry_device.cpp
@@ -1,0 +1,228 @@
+// ======================================================================== //
+// Copyright 2009-2018 Intel Corporation                                    //
+//                                                                          //
+// Licensed under the Apache License, Version 2.0 (the "License");          //
+// you may not use this file except in compliance with the License.         //
+// You may obtain a copy of the License at                                  //
+//                                                                          //
+//     http://www.apache.org/licenses/LICENSE-2.0                           //
+//                                                                          //
+// Unless required by applicable law or agreed to in writing, software      //
+// distributed under the License is distributed on an "AS IS" BASIS,        //
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. //
+// See the License for the specific language governing permissions and      //
+// limitations under the License.                                           //
+// ======================================================================== //
+
+#include "../common/tutorial/tutorial_device.h"
+#include "../common/tutorial/optics.h"
+#include "../common/math/random_sampler.h"
+
+namespace embree {
+
+/* scene data */
+RTCScene  g_scene  = nullptr;
+
+#define NUM_POINTS 512
+Vec3fa point_colors[NUM_POINTS];
+
+/* add point geometry */
+void addPoints (RTCScene scene, RTCGeometryType gtype, const Vec3fa& pos)
+{
+  RandomSampler rng;
+  RandomSampler_init(rng, 42);
+
+#define COORD RandomSampler_get1D(rng) * 4.f - 2.f
+#define RADIUS RandomSampler_get1D(rng) * 0.13f + 0.02f
+#define COLOR RandomSampler_get1D(rng)
+#define NORMAL RandomSampler_get1D(rng) * 2.f - 1.f
+
+  RTCGeometry geom = rtcNewGeometry(g_device, gtype);
+  Vec4f *point_vertices = (Vec4f*)rtcSetNewGeometryBuffer(geom,RTC_BUFFER_TYPE_VERTEX, 0, RTC_FORMAT_FLOAT4, sizeof(Vec4f), NUM_POINTS);
+
+  for (int i = 0; i < NUM_POINTS; i++) {
+    float vtx[4];
+    vtx[0] = COORD;
+    vtx[1] = COORD;
+    vtx[2] = COORD;
+    vtx[3] = RADIUS;
+    float color[3];
+    color[0] = COLOR;
+    color[1] = COLOR;
+    color[2] = COLOR;
+    point_vertices[i] = Vec4f(pos) + Vec4f(vtx[0], vtx[1], vtx[2], vtx[3]);
+    point_colors[i] = Vec3fa(color[0], color[1], color[2]);
+  }
+
+  if (gtype == RTC_GEOMETRY_TYPE_ORIENTED_DISC_POINT) {
+    Vec3fa *point_normals = (Vec3fa*)rtcSetNewGeometryBuffer(geom,RTC_BUFFER_TYPE_NORMAL, 0, RTC_FORMAT_FLOAT3, sizeof(Vec3fa), NUM_POINTS);
+    for (int i = 0; i < NUM_POINTS; i++) {
+      float normal[3];
+      normal[0] = NORMAL;
+      normal[1] = NORMAL;
+      normal[2] = NORMAL;
+      point_normals[i] = Vec3fa(normal[0], normal[1], normal[2]);
+      point_normals[i] = normalize(point_normals[i]);
+    }
+  }
+
+  rtcCommitGeometry(geom);
+  rtcAttachGeometry(scene,geom);
+  rtcReleaseGeometry(geom);
+}
+
+/* adds a ground plane to the scene */
+unsigned int addGroundPlane (RTCScene scene_i)
+{
+  /* create a triangulated plane with 2 triangles and 4 vertices */
+  RTCGeometry geom = rtcNewGeometry (g_device, RTC_GEOMETRY_TYPE_TRIANGLE);
+
+  /* set vertices */
+  Vertex* vertices = (Vertex*) rtcSetNewGeometryBuffer(geom, RTC_BUFFER_TYPE_VERTEX, 0, RTC_FORMAT_FLOAT3, sizeof(Vertex), 4);
+  vertices[0].x = -10; vertices[0].y = -2; vertices[0].z = -10;
+  vertices[1].x = -10; vertices[1].y = -2; vertices[1].z = +10;
+  vertices[2].x = +10; vertices[2].y = -2; vertices[2].z = -10;
+  vertices[3].x = +10; vertices[3].y = -2; vertices[3].z = +10;
+
+  /* set triangles */
+  Triangle* triangles = (Triangle*) rtcSetNewGeometryBuffer(geom, RTC_BUFFER_TYPE_INDEX, 0, RTC_FORMAT_UINT3, sizeof(Triangle), 2);
+  triangles[0].v0 = 0; triangles[0].v1 = 1; triangles[0].v2 = 2;
+  triangles[1].v0 = 1; triangles[1].v1 = 3; triangles[1].v2 = 2;
+
+  rtcCommitGeometry(geom);
+  unsigned int geomID = rtcAttachGeometry(scene_i,geom);
+  rtcReleaseGeometry(geom);
+  return geomID;
+}
+
+/* called by the C++ code for initialization */
+extern "C" void device_init (char* cfg)
+{
+  /* create scene */
+  g_scene = rtcNewScene(g_device);
+
+  /* add ground plane */
+  addGroundPlane(g_scene);
+
+  /* add curve */
+  addPoints(g_scene, RTC_GEOMETRY_TYPE_SPHERE_POINT,        Vec3fa( 0.0f, 0.0f, 0.0f));
+  addPoints(g_scene, RTC_GEOMETRY_TYPE_DISC_POINT,          Vec3fa( 5.0f, 0.0f, 0.0f));
+  addPoints(g_scene, RTC_GEOMETRY_TYPE_ORIENTED_DISC_POINT, Vec3fa(-5.0f, 0.0f, 0.0f));
+
+  /* commit changes to scene */
+  rtcCommitScene (g_scene);
+
+  /* set start render mode */
+  renderTile = renderTileStandard;
+  key_pressed_handler = device_key_pressed_default;
+}
+
+/* task that renders a single screen tile */
+Vec3fa renderPixelStandard(float x, float y, const ISPCCamera& camera, RayStats& stats)
+{
+  RTCIntersectContext context;
+  rtcInitIntersectContext(&context);
+
+  /* initialize ray */
+  Ray ray(Vec3fa(camera.xfm.p), Vec3fa(normalize(x*camera.xfm.l.vx + y*camera.xfm.l.vy + camera.xfm.l.vz)), 0.0f, inf);
+
+  /* intersect ray with scene */
+  rtcIntersect1(g_scene,&context,RTCRayHit_(ray));
+  RayStats_addRay(stats);
+
+  /* shade pixels */
+  Vec3fa color = Vec3fa(0.0f);
+  if (ray.geomID != RTC_INVALID_GEOMETRY_ID)
+  {
+    /* interpolate diffuse color */
+    Vec3fa diffuse = point_colors[ray.geomID ? ray.primID : 0];
+
+    /* calculate smooth shading normal */
+    Vec3fa Ng = normalize(ray.Ng);
+    color = color + diffuse*0.5f;
+    Vec3fa lightDir = normalize(Vec3fa(-1,-1,-1));
+
+    /* initialize shadow ray */
+    Ray shadow(ray.org + ray.tfar*ray.dir, neg(lightDir), 0.001f, inf, 0.0f);
+
+    /* trace shadow ray */
+    rtcOccluded1(g_scene,&context,RTCRay_(shadow));
+    RayStats_addShadowRay(stats);
+
+    /* add light contribution */
+    if (shadow.tfar >= 0.0f) {
+      Vec3fa r = normalize(reflect(ray.dir,Ng));
+      float s = pow(clamp(dot(r,lightDir),0.0f,1.0f),10.0f);
+      float d = clamp(-dot(lightDir,Ng),0.0f,1.0f);
+      color = color + diffuse*d + 0.5f*Vec3fa(s);
+    }
+  }
+  return color;
+}
+
+/* renders a single screen tile */
+void renderTileStandard(int taskIndex,
+                        int threadIndex,
+                        int* pixels,
+                        const unsigned int width,
+                        const unsigned int height,
+                        const float time,
+                        const ISPCCamera& camera,
+                        const int numTilesX,
+                        const int numTilesY)
+{
+  const unsigned int tileY = taskIndex / numTilesX;
+  const unsigned int tileX = taskIndex - tileY * numTilesX;
+  const unsigned int x0 = tileX * TILE_SIZE_X;
+  const unsigned int x1 = min(x0+TILE_SIZE_X,width);
+  const unsigned int y0 = tileY * TILE_SIZE_Y;
+  const unsigned int y1 = min(y0+TILE_SIZE_Y,height);
+
+  for (unsigned int y=y0; y<y1; y++) for (unsigned int x=x0; x<x1; x++)
+  {
+    /* calculate pixel color */
+    Vec3fa color = renderPixelStandard((float)x,(float)y,camera,g_stats[threadIndex]);
+
+    /* write color to framebuffer */
+    unsigned int r = (unsigned int) (255.0f * clamp(color.x,0.0f,1.0f));
+    unsigned int g = (unsigned int) (255.0f * clamp(color.y,0.0f,1.0f));
+    unsigned int b = (unsigned int) (255.0f * clamp(color.z,0.0f,1.0f));
+    pixels[y*width+x] = (b << 16) + (g << 8) + r;
+  }
+}
+
+/* task that renders a single screen tile */
+void renderTileTask (int taskIndex, int threadIndex, int* pixels,
+                         const unsigned int width,
+                         const unsigned int height,
+                         const float time,
+                         const ISPCCamera& camera,
+                         const int numTilesX,
+                         const int numTilesY)
+{
+  renderTile(taskIndex,threadIndex,pixels,width,height,time,camera,numTilesX,numTilesY);
+}
+
+/* called by the C++ code to render */
+extern "C" void device_render (int* pixels,
+                    const unsigned int width,
+                    const unsigned int height,
+                    const float time,
+                    const ISPCCamera& camera)
+{
+  const int numTilesX = (width +TILE_SIZE_X-1)/TILE_SIZE_X;
+  const int numTilesY = (height+TILE_SIZE_Y-1)/TILE_SIZE_Y;
+  parallel_for(size_t(0),size_t(numTilesX*numTilesY),[&](const range<size_t>& range) {
+    const int threadIndex = (int)TaskScheduler::threadIndex();
+    for (size_t i=range.begin(); i<range.end(); i++)
+      renderTileTask((int)i,threadIndex,pixels,width,height,time,camera,numTilesX,numTilesY);
+  });
+}
+
+/* called by the C++ code for cleanup */
+extern "C" void device_cleanup ()
+{
+  rtcReleaseScene (g_scene); g_scene = nullptr;
+}
+
+} // namespace embree